+++ app/Http/Controllers/EducationDegreesController.php
@@ -6,6 +6,7 @@ use App\EducationDegree;
 use App\Faculty;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Redirect;
 
 class EducationDegreesController extends Controller
 {
@@ -25,7 +26,7 @@ class EducationDegreesController extends Controller
         if (Auth::user() == null)
             return redirect('/login');
         if (!Auth::user()->is_admin)
-            return redirect('/403');
+            abort(403);
         $this->validate($request, [
             'name' => 'required',
             'description' => 'required'
@@ -43,6 +44,8 @@ class EducationDegreesController extends Controller
     {
         $edu_deg = EducationDegree::all();
         $deg = EducationDegree::all()->where('name', '=', $name)->first();
+        if($deg == null)
+            abort(404);
         $faculties = Faculty::all()->where('education_degree_id', '=', $deg->id);
         return view("pages/edu_deg/show")
             ->with('title', "Education Degree - Admission")
@@ -55,8 +58,10 @@ class EducationDegreesController extends Controller
         if (Auth::user() == null)
             return redirect('/login');
         if (!Auth::user()->is_admin)
-            return redirect('/403');
+            abort(403);
         $edu_deg = EducationDegree::find($id);
+        if ($edu_deg==null)
+            abort(404);
         return view("pages/admin/edu_deg_edit")->with('title', "Education Degree - Admission")->with("edu_deg", $edu_deg);
     }
 
@@ -65,13 +70,16 @@ class EducationDegreesController extends Controller
         if (Auth::user() == null)
             return redirect('/login');
         if (!Auth::user()->is_admin)
-            return redirect('/403');
+            abort(403);
         $this->validate($request, [
             'name' => 'required',
             'description' => 'required'
         ]);
 
         $edu_deg = EducationDegree::find($id);
+
+        if($edu_deg==null)
+            abort(404);
 
         $edu_deg->name = $request->input('name');
         $edu_deg->description = $request->input('description');
@@ -86,8 +94,10 @@ class EducationDegreesController extends Controller
         if (Auth::user() == null)
             return redirect('/login');
         if (!Auth::user()->is_admin)
-            return redirect('/403');
+            abort(403);
         $edu_deg = EducationDegree::find($id);
+        if($edu_deg==null)
+            abort(404);
         $edu_deg->delete();
         return redirect('/admin/education_degrees')->with('success', 'Education updated');
     }
